requirements.txt
dash==2.0.0 
numpy==1.21.2
pandas==1.3.4 
plotly==5.3.1
pytables==3.6.1
python==3.9.7
python-dateutil==2.8.2
scikit-learn==1.0.1
tqdm==4.62.3